second attempt at making modules
