+++ rust/exo_rs/python/.gitignore
@@ -1,0 +1,4 @@
+*.so
+*.pyd
+*.dll
+*.dylib
