+++ rust/exo_rs/python/exo_rs/__init__.py
@@ -1,0 +1,4 @@
+# This file is automatically generated by pyo3_stub_gen
+# ruff: noqa: F401
+
+__all__ = []
