--- rust/exo_rs/python/exo_rs/__init__.pyi
@@ -1,12 +1,0 @@
-# This file is automatically generated by pyo3_stub_gen
-# ruff: noqa: E501, F401, F403, F405
-
-from . import ident
-from . import networking
-from . import pidfile
-
-__all__ = [
-    "ident",
-    "networking",
-    "pidfile",
-]
