+++ rust/exo_rs/python/exo_rs/ident/__init__.py
@@ -1,0 +1,8 @@
+# This file is automatically generated by pyo3_stub_gen
+# ruff: noqa: F401
+
+from exo_rs._core import Keypair
+
+__all__ = [
+    "Keypair",
+]
