--- rust/exo_rs/python/exo_rs/networking/__init__.pyi
@@ -1,94 +1,0 @@
-# This file is automatically generated by pyo3_stub_gen
-# ruff: noqa: E501, F401, F403, F405
-
-import builtins
-from exo_rs import ident
-import typing
-
-__all__ = [
-    "AllQueuesFullError",
-    "FromSwarm",
-    "MessageTooLargeError",
-    "NetworkingHandle",
-    "NoPeersSubscribedToTopicError",
-]
-
-@typing.final
-class AllQueuesFullError(builtins.Exception):
-    def __new__(cls, *args: typing.Any) -> AllQueuesFullError: ...
-    def __repr__(self) -> builtins.str: ...
-    def __str__(self) -> builtins.str: ...
-
-class FromSwarm:
-    @typing.final
-    class Connection(FromSwarm):
-        __match_args__ = (
-            "peer_id",
-            "connected",
-        )
-        @property
-        def peer_id(self) -> builtins.str: ...
-        @property
-        def connected(self) -> builtins.bool: ...
-        def __new__(
-            cls, peer_id: builtins.str, connected: builtins.bool
-        ) -> FromSwarm.Connection: ...
-
-    @typing.final
-    class Message(FromSwarm):
-        __match_args__ = (
-            "origin",
-            "topic",
-            "data",
-        )
-        @property
-        def origin(self) -> builtins.str: ...
-        @property
-        def topic(self) -> builtins.str: ...
-        @property
-        def data(self) -> bytes: ...
-        def __new__(
-            cls, origin: builtins.str, topic: builtins.str, data: bytes
-        ) -> FromSwarm.Message: ...
-
-    ...
-
-@typing.final
-class MessageTooLargeError(builtins.Exception):
-    def __new__(cls, *args: typing.Any) -> MessageTooLargeError: ...
-    def __repr__(self) -> builtins.str: ...
-    def __str__(self) -> builtins.str: ...
-
-@typing.final
-class NetworkingHandle:
-    def __new__(
-        cls,
-        identity: ident.Keypair,
-        bootstrap_peers: typing.Sequence[builtins.str],
-        listen_port: builtins.int,
-    ) -> NetworkingHandle: ...
-    def recv(self) -> typing.Awaitable[FromSwarm]: ...
-    async def gossipsub_subscribe(self, topic: builtins.str) -> builtins.bool:
-        r"""
-        Subscribe to a `GossipSub` topic.
-
-        Returns `True` if the subscription worked. Returns `False` if we were already subscribed.
-        """
-    async def gossipsub_unsubscribe(self, topic: builtins.str) -> builtins.bool:
-        r"""
-        Unsubscribes from a `GossipSub` topic.
-
-        Returns `True` if we were subscribed to this topic. Returns `False` if we were not subscribed.
-        """
-    async def gossipsub_publish(self, topic: builtins.str, data: bytes) -> None:
-        r"""
-        Publishes a message with multiple topics to the `GossipSub` network.
-
-        If no peers are found that subscribe to this topic, throws `NoPeersSubscribedToTopicError` exception.
-        """
-
-@typing.final
-class NoPeersSubscribedToTopicError(builtins.Exception):
-    def __new__(cls, *args: typing.Any) -> NoPeersSubscribedToTopicError: ...
-    def __repr__(self) -> builtins.str: ...
-    def __str__(self) -> builtins.str: ...
