+++ rust/exo_rs/python/exo_rs/pidfile/__init__.py
@@ -1,0 +1,9 @@
+# This file is automatically generated by pyo3_stub_gen
+# ruff: noqa: F401
+
+from exo_rs._core import Pidfile, PidfileError
+
+__all__ = [
+    "Pidfile",
+    "PidfileError",
+]
