--- rust/exo_rs/python/exo_rs/pidfile/__init__.pyi
@@ -1,70 +1,0 @@
-# This file is automatically generated by pyo3_stub_gen
-# ruff: noqa: E501, F401, F403, F405
-
-import builtins
-import os
-import pathlib
-import typing
-
-__all__ = [
-    "Pidfile",
-    "PidfileError",
-]
-
-@typing.final
-class Pidfile:
-    r"""
-    A PID file protected with a lock.
-
-    An instance of `Pidfile` can be used to manage a PID file: create it,
-    lock it, detect already running daemons. It is backed by [`pidfile`]
-    functions of `libbsd`/`libutil` which use `flopen` to lock the PID
-    file.
-
-    When a PID file is created, the process ID of the current process is
-    *not* written there, making it possible to lock the PID file before
-    forking and only write the ID of the forked process when it is ready.
-
-    The PID file is deleted automatically when the `Pidfile` comes out of
-    the scope. To close the PID file without deleting it, for example, in
-    the parent process of a forked daemon, call `close()`.
-
-    [`exit`]: https://doc.rust-lang.org/std/process/fn.exit.html
-    [`pidfile`]: https://linux.die.net/man/3/pidfile
-    [`daemon`(3)]: https://linux.die.net/man/3/daemon
-    """
-    def __new__(
-        cls, path: builtins.str | os.PathLike | pathlib.Path, mode: builtins.int
-    ) -> Pidfile:
-        r"""
-        Creates a new PID file and locks it.
-
-        If the PID file cannot be locked, returns `PidfileError::AlreadyRunning` with
-        a PID of the already running process, or `None` if no PID has been written to
-        the PID file yet.
-        """
-    def write(self) -> None:
-        r"""
-        Writes the current process ID to the PID file.
-
-        The file is truncated before writing.
-        """
-    def as_raw_fd(self) -> builtins.int:
-        r"""
-        Extracts the raw file descriptor.
-
-        This function is typically used to **borrow** an owned file descriptor.
-        When used in this way, this method does **not** pass ownership of the
-        raw file descriptor to the caller, and the file descriptor is only
-        guaranteed to be valid while the original object has not yet been
-        destroyed.
-        """
-    def close(self) -> None:
-        r"""
-        Closes the PID file and releases associated resources.
-        """
-
-@typing.final
-class PidfileError(builtins.Exception):
-    def __repr__(self) -> builtins.str: ...
-    def __str__(self) -> builtins.str: ...
