+++ rust/exo_rs/tests/test_python.py
@@ -1,14 +1,15 @@
 import asyncio
+from contextlib import suppress
 
 import pytest
 from _pytest.capture import CaptureFixture
-from exo_rs import (
-    Keypair,
+from exo_rs.ident import Keypair
+from exo_rs.networking import (
+    FromSwarm,
     NetworkingHandle,
     NoPeersSubscribedToTopicError,
-    Pidfile,
-    FromSwarm,
 )
+from exo_rs.pidfile import Pidfile
 
 
 @pytest.mark.asyncio
@@ -16,16 +17,21 @@ async def test_sleep_on_multiple_items() -> None:
     print("PYTHON: starting handle")
     h = NetworkingHandle(Keypair.generate(), [], 0)
 
-    rt = asyncio.create_task(_await_recv(h))
+    recv_task = asyncio.create_task(_await_recv(h))
 
-    # sleep for 4 ticks
-    for i in range(4):
-        await asyncio.sleep(1)
+    try:
+        # sleep for 4 ticks
+        for _ in range(4):
+            await asyncio.sleep(1)
 
-        try:
-            await h.gossipsub_publish("topic", b"somehting or other")
-        except NoPeersSubscribedToTopicError as e:
-            print("caught it", e)
+            try:
+                await h.gossipsub_publish("topic", b"somehting or other")
+            except NoPeersSubscribedToTopicError as e:
+                print("caught it", e)
+    finally:
+        recv_task.cancel()
+        with suppress(asyncio.CancelledError):
+            await recv_task
 
 
 def test_pidfile(capsys: CaptureFixture[str]):
@@ -46,4 +52,5 @@ async def _await_recv(h: NetworkingHandle):
 
 
 def scoped_lock_file():
-    a = Pidfile("/tmp/lock.pid", 0o0600)
+    lock_file = Pidfile("/tmp/lock.pid", 0o0600)
+    lock_file.close()
