+++ src/exo/routing/connection_message.py
@@ -1,4 +1,4 @@
-from exo_rs import FromSwarm
+from exo_rs.networking import FromSwarm
 
 from exo.shared.types.common import NodeId
 from exo.utils.pydantic_ext import FrozenModel
